refactor: move agentHeader.js and agentHeader.scss to header depository
