+++ src/agent/header/agentHeader.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "src/variables";
 .agentHeader{
   margin:1rem 0;
 }
